# (PC-12040) Fix adage query performance issue