--- a/api/tests/routes/adage/v1/get_educational_institution_test.py
+++ b/api/tests/routes/adage/v1/get_educational_institution_test.py
@@ -6,6 +6,7 @@ from pcapi.core.educational.factories import EducationalInstitutionFactory
 from pcapi.core.educational.factories import EducationalRedactorFactory
 from pcapi.core.educational.factories import EducationalYearFactory
 from pcapi.core.offers.utils import offer_webapp_link
+from pcapi.core.testing import assert_num_queries
 from pcapi.utils.date import format_into_utc_date
 
 from tests.conftest import TestClient
@@ -108,6 +109,53 @@ class Returns200Test:
                 }
             ],
         }
+
+    def test_get_educational_institution_num_queries(self, app) -> None:
+        redactor = EducationalRedactorFactory()
+        educational_year = EducationalYearFactory()
+        educational_institution = EducationalInstitutionFactory()
+        EducationalDepositFactory(
+            educationalInstitution=educational_institution,
+            educationalYear=educational_year,
+            amount=200000,
+            isFinal=True,
+        )
+        EducationalInstitutionFactory()
+        EducationalBookingFactory.create_batch(
+            10,
+            educationalBooking__educationalRedactor=redactor,
+            educationalBooking__educationalYear=educational_year,
+            educationalBooking__educationalInstitution=educational_institution,
+        )
+        other_educational_year = EducationalYearFactory(adageId="toto")
+        other_educational_institution = EducationalInstitutionFactory(institutionId="tata")
+        EducationalBookingFactory(educationalBooking__educationalYear=other_educational_year)
+        EducationalBookingFactory(educationalBooking__educationalInstitution=other_educational_institution)
+        EducationalDepositFactory(
+            educationalInstitution=educational_institution,
+            educationalYear=other_educational_year,
+            amount=2000,
+            isFinal=False,
+        )
+        EducationalDepositFactory(
+            educationalInstitution=other_educational_institution,
+            educationalYear=educational_year,
+            amount=2000,
+            isFinal=False,
+        )
+
+        client = TestClient(app.test_client()).with_eac_token()
+        adage_id = educational_year.adageId
+        institution_id = educational_institution.institutionId
+
+        n_queries = 0
+        n_queries += 1  # Check for educational institution
+        n_queries += 1  # Get needed data
+        n_queries += 1  # Get deposit
+        n_queries += 1  # Check webapp_v2 feature toggle
+
+        with assert_num_queries(n_queries):
+            client.get(f"/adage/v1/years/{adage_id}/educational_institution/{institution_id}")
 
     def test_get_educational_institution_without_deposit(self, app) -> None:
         educational_year = EducationalYearFactory()